break another circdep with globToRegExp
