+++ src/core/env.ts
@@ -1,11 +1,8 @@
 /*
-* env.ts
-*
-* Copyright (C) 2020-2022 Posit Software, PBC
-*
-*/
-
-import { expandPath } from "./path.ts";
+ * env.ts
+ *
+ * Copyright (C) 2020-2022 Posit Software, PBC
+ */
 
 export function getenv(name: string, defaultValue?: string) {
   const value = Deno.env.get(name);
@@ -43,28 +40,5 @@ export function withPath(
       modifiedPaths.push(...paths.prepend);
     }
     return modifiedPaths.join(delimiter);
-  }
-}
-
-export function suggestUserBinPaths() {
-  if (Deno.build.os !== "windows") {
-    // List of paths that we consider bin paths
-    // in priority order (expanded and not)
-    const possiblePaths = [
-      "/usr/local/bin",
-      "~/.local/bin",
-      "~/bin",
-    ];
-
-    // Read the user path
-    const pathRaw = Deno.env.get("PATH");
-    const paths: string[] = pathRaw ? pathRaw.split(":") : [];
-
-    // Filter the above list by what is in the user path
-    return possiblePaths.filter((path) => {
-      return paths.includes(path) || paths.includes(expandPath(path));
-    });
-  } else {
-    throw new Error("suggestUserBinPaths not currently supported on Windows");
   }
 }
