+++ src/core/path.ts
@@ -299,3 +299,27 @@ export function normalizePath(path: string | URL): string {
   // letters, make those uppercase
   return file.replace(/^\w:\\/, (m) => m[0].toUpperCase() + ":\\");
 }
+
+// Moved here from env.ts to avoid circular dependency
+export function suggestUserBinPaths() {
+  if (Deno.build.os !== "windows") {
+    // List of paths that we consider bin paths
+    // in priority order (expanded and not)
+    const possiblePaths = [
+      "/usr/local/bin",
+      "~/.local/bin",
+      "~/bin",
+    ];
+
+    // Read the user path
+    const pathRaw = Deno.env.get("PATH");
+    const paths: string[] = pathRaw ? pathRaw.split(":") : [];
+
+    // Filter the above list by what is in the user path
+    return possiblePaths.filter((path) => {
+      return paths.includes(path) || paths.includes(expandPath(path));
+    });
+  } else {
+    throw new Error("suggestUserBinPaths not currently supported on Windows");
+  }
+}
